@@ -1,6 +1,6 @@
 module github.com/telekom/das-schiff-network-operator
 
-go 1.19
+go 1.20
 
 require (
 	github.com/cilium/ebpf v0.9.1
@@ -12,7 +12,8 @@ require (
 	github.com/onsi/gomega v1.27.10
 	github.com/prometheus/client_golang v1.16.0
 	github.com/vishvananda/netlink v1.1.1-0.20211129163951-9ada19101fc5
-	go.uber.org/mock v0.2.0
+	go.uber.org/mock v0.3.0
+	go4.org/netipx v0.0.0-20231129151722-fdeea329fbba
 	golang.org/x/sys v0.13.0
 	gopkg.in/yaml.v2 v2.4.0
 	k8s.io/api v0.28.3
@@ -77,7 +78,6 @@ require (
 	k8s.io/component-base v0.28.3 // indirect
 	k8s.io/klog/v2 v2.100.1 // indirect
 	k8s.io/kube-openapi v0.0.0-20230717233707-2695361300d9 // indirect
-	k8s.io/kubernetes v1.28.4
 	sigs.k8s.io/json v0.0.0-20221116044647-bc3834ca7abd // indirect
 	sigs.k8s.io/structured-merge-diff/v4 v4.2.3 // indirect
 	sigs.k8s.io/yaml v1.3.0 // indirect
