@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/telekom/das-schiff-network-operator/pkg/healthcheck (interfaces: FRRInterface,TCPDialerInterface)
-
+//
+// Generated by this command:
+//
+//	mockgen -destination ./mock/mock_healthcheck.go . FRRInterface,TCPDialerInterface
+//
 // Package mock_healthcheck is a generated GoMock package.
 package mock_healthcheck
 
@@ -83,7 +87,7 @@ func (m *MockTCPDialerInterface) Dial(arg0, arg1 string) (net.Conn, error) {
 }
 
 // Dial indicates an expected call of Dial.
-func (mr *MockTCPDialerInterfaceMockRecorder) Dial(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockTCPDialerInterfaceMockRecorder) Dial(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Dial", reflect.TypeOf((*MockTCPDialerInterface)(nil).Dial), arg0, arg1)
 }
